api/requirements.txt
azure-functions==1.11.2
certifi==2022.6.15
charset-normalizer==2.1.0
greenlet==1.1.2
idna==3.3
numpy==1.23.1
pandas==1.4.3
pyodbc==4.0.32
pyTelegramBotAPI==4.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
tabulate==0.8.10
urllib3==1.26.10
